chatbot_project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
datasets==3.5.0
dill==0.3.8
Django==5.2
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langsmith==0.3.32
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
propcache==0.3.1
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
